requirements.txt
pandas
numpy
pymssql
python-dotenv
